datepicker: refactor code and add flex prop (#35)
* datepicker: refactor code and add flex prop

* date_picker: remove extranous comment

* nav_bar: rename file

* input_icon: rename file

* app.routes: update path

* date_picker: add style prop and icon render


---------
