+++ src/components/date_picker.tsx
@@ -3,19 +3,33 @@ import RNDateTimePicker, {
     DateTimePickerEvent,
 } from "@react-native-community/datetimepicker";
 import React, { Dispatch, SetStateAction, useState } from "react";
-import { Keyboard, View } from "react-native";
+import { Keyboard, StyleProp, StyleSheet, View, ViewStyle } from "react-native";
 import colors from "../pallete";
 import InputIcon from "./input_icon";
 
-export default function DatePicker({
-    text,
-    textState,
-    icon,
-}: {
+export interface DatePickerProps {
     text: string;
+    style: StyleProp<ViewStyle>;
     textState: Dispatch<SetStateAction<string>>;
     icon: boolean;
-}) {
+}
+
+const styles = StyleSheet.create({
+    input: {
+        height: 40,
+        borderBottomWidth: 1,
+        borderColor: colors.black_400,
+        paddingLeft: 15,
+        borderRadius: 0,
+        minWidth: 100,
+        color: colors.black_500,
+        fontSize: 15,
+    },
+});
+
+export default function DatePicker(props: DatePickerProps) {
+    const { text, textState, style, icon } = props;
+
     const [date, setDate] = useState(new Date(1598051730000));
     const [show, setShow] = useState(false);
     const dataMax = new Date();
@@ -38,17 +52,7 @@ export default function DatePicker({
     return (
         <View>
             <InputIcon
-                style={{
-                    height: 40,
-                    borderBottomWidth: 1,
-                    borderColor: colors.black_400,
-                    paddingLeft: 15,
-                    // paddingRight: 25,
-                    borderRadius: 0,
-                    minWidth: 100,
-                    color: colors.black_500,
-                    fontSize: 15,
-                }}
+                style={[styles.input, style]}
                 onFocus={() => {
                     Keyboard.dismiss();
                     openDatePicker();
@@ -58,11 +62,14 @@ export default function DatePicker({
                 placeholder="AAAA-MM-DD"
                 keyboardType="numeric"
                 Icon={
-                    <FontAwesome
-                        name="calendar"
-                        size={20}
-                        color={colors.black_400}
-                    />
+                    (icon && (
+                        <FontAwesome
+                            name="calendar"
+                            size={20}
+                            color={colors.black_400}
+                        />
+                    )) ||
+                    undefined
                 }
             />
